+++ src/vm/int_array.h
@@ -12,8 +12,8 @@ public:
   static IntArray *Copy(const IntArray *mem);
 
   virtual ~IntArray();
-  virtual void Write(int addr, const numeric::Number &data) = 0;
-  virtual numeric::Number Read(int addr) = 0;
+  virtual void Write(uint64_t addr, const numeric::Number &data) = 0;
+  virtual numeric::Number Read(uint64_t addr) = 0;
   virtual long long GetLength() const = 0;
   virtual const iroha::NumericWidth *GetWidth() const = 0;
   virtual int GetAddressWidth() const = 0;
